--- gamertrack-application/src/main/java/com/gepardec/GameDoesNotExistExceptionMapper.java
@@ -1,17 +1,0 @@
-package com.gepardec;
-
-import com.gepardec.service.GameDoesNotExistException;
-import jakarta.ws.rs.core.Response;
-import jakarta.ws.rs.core.Response.Status;
-import jakarta.ws.rs.ext.ExceptionMapper;
-import jakarta.ws.rs.ext.Provider;
-
-@Provider
-public class GameDoesNotExistExceptionMapper implements ExceptionMapper<GameDoesNotExistException> {
-
-  @Override
-  public Response toResponse(GameDoesNotExistException exception) {
-
-    return Response.status(Status.NOT_FOUND).entity(exception.getMessage()).build();
-  }
-}
