--- gamertrack-application/src/main/java/com/gepardec/games/CreateGameCommand.java
@@ -1,7 +1,0 @@
-package com.gepardec.games;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record CreateGameCommand(@NotBlank String title, String rules) {
-
-}
