--- gamertrack-application/src/main/java/com/gepardec/games/GameDto.java
@@ -1,10 +1,0 @@
-package com.gepardec.games;
-
-import com.gepardec.model.Game;
-
-public record GameDto(Long id , String name, String rules) {
-
-  public GameDto(Game Game) {
-    this(Game.getId(), Game.getName(), Game.getRules());
-  }
-}
