--- gamertrack-application/src/main/java/com/gepardec/games/UpdateGameCommand.java
@@ -1,5 +1,0 @@
-package com.gepardec.games;
-
-public record UpdateGameCommand(Long id, String title, String rules) {
-
-}
