+++ gamertrack-application/src/main/java/com/gepardec/rest/model/command/CreateGameCommand.java
@@ -1,0 +1,7 @@
+package com.gepardec.rest.model.command;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record CreateGameCommand(@NotBlank String title, String rules) {
+
+}
