+++ gamertrack-application/src/main/java/com/gepardec/rest/model/command/CreateUserCommand.java
@@ -1,0 +1,29 @@
+package com.gepardec.rest.model.command;
+
+public class CreateUserCommand {
+    private String firstname;
+    private String lastname;
+
+    public CreateUserCommand() {}
+
+    public CreateUserCommand(String firstname, String lastname) {
+        this.firstname = firstname;
+        this.lastname = lastname;
+    }
+
+    public String getFirstname() {
+        return firstname;
+    }
+
+    public void setFirstname(String firstname) {
+        this.firstname = firstname;
+    }
+
+    public String getLastname() {
+        return lastname;
+    }
+
+    public void setLastname(String lastname) {
+        this.lastname = lastname;
+    }
+}
