+++ gamertrack-application/src/main/java/com/gepardec/rest/model/command/UpdateGameCommand.java
@@ -1,0 +1,5 @@
+package com.gepardec.rest.model.command;
+
+public record UpdateGameCommand(Long id, String title, String rules) {
+
+}
