--- gamertrack-application/src/main/java/com/gepardec/users/cmds/CreateUserCommand.java
@@ -1,29 +1,0 @@
-package com.gepardec.users.cmds;
-
-public class CreateUserCommand {
-    private String firstname;
-    private String lastname;
-
-    public CreateUserCommand() {}
-
-    public CreateUserCommand(String firstname, String lastname) {
-        this.firstname = firstname;
-        this.lastname = lastname;
-    }
-
-    public String getFirstname() {
-        return firstname;
-    }
-
-    public void setFirstname(String firstname) {
-        this.firstname = firstname;
-    }
-
-    public String getLastname() {
-        return lastname;
-    }
-
-    public void setLastname(String lastname) {
-        this.lastname = lastname;
-    }
-}
