--- gamertrack-application/src/main/java/com/gepardec/users/dto/UserDto.java
@@ -1,11 +1,0 @@
-package com.gepardec.users.dto;
-
-
-import com.gepardec.model.User;
-import jakarta.validation.constraints.NotEmpty;
-
-public record UserDto (@NotEmpty String firstname, @NotEmpty String lastname) {
-    public UserDto(User user){
-        this(user.getFirstname(),user.getLastname());
-    }
-}
