+++ gamertrack-db/pom.xml
@@ -10,7 +10,7 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>gamertrack-application</artifactId>
+    <artifactId>gamertrack-db</artifactId>
 
     <properties>
         <maven.compiler.source>23</maven.compiler.source>
@@ -24,23 +24,8 @@
             <artifactId>gamertrack-domain</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>com.gepardec</groupId>
-            <artifactId>gamertrack-service</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>jakarta.servlet</groupId>
-            <artifactId>jakarta.servlet-api</artifactId>
-            <version>6.1.0</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>jakarta.ws.rs</groupId>
-            <artifactId>jakarta.ws.rs-api</artifactId>
-            <version>4.0.0</version>
-            <scope>provided</scope>
-        </dependency>
+
+
     </dependencies>
 
 </project>
