--- gamertrack-infrastructure/src/main/java/com/gepardec/service/GameAlreadyExistsException.java
@@ -1,9 +1,0 @@
-package com.gepardec.service;
-
-public class GameAlreadyExistsException extends RuntimeException {
-
-  public GameAlreadyExistsException(String message) {
-    super(message);
-  }
-
-}
