--- gamertrack-infrastructure/src/main/java/com/gepardec/service/GameDoesNotExistException.java
@@ -1,8 +1,0 @@
-package com.gepardec.service;
-
-public class GameDoesNotExistException extends RuntimeException {
-  public GameDoesNotExistException(String message) {
-    super(message);
-  }
-
-}
