+++ gamertrack-service/src/main/java/com/gepardec/impl/GameAlreadyExistsException.java
@@ -1,0 +1,9 @@
+package com.gepardec.impl;
+
+public class GameAlreadyExistsException extends RuntimeException {
+
+  public GameAlreadyExistsException(String message) {
+    super(message);
+  }
+
+}
