+++ gamertrack-service/src/main/java/com/gepardec/impl/GameDoesNotExistException.java
@@ -1,0 +1,8 @@
+package com.gepardec.impl;
+
+public class GameDoesNotExistException extends RuntimeException {
+  public GameDoesNotExistException(String message) {
+    super(message);
+  }
+
+}
